Automatically detect and set zero terminal SNR
